+++ test/net/sf/persism/TestStackOverflow.java
@@ -154,82 +154,83 @@ public class TestStackOverflow extends TestCase {
         // Posts IN
         sql = """
                 SELECT [Id], [AcceptedAnswerId], [AnswerCount], [Body], [ClosedDate], [CommentCount], [CommunityOwnedDate],
-                 [CreationDate], [FavoriteCount], [LastActivityDate], [LastEditDate], [LastEditorDisplayName], [LastEditorUserId], 
-                 [OwnerUserId], [ParentId], [PostTypeId], [Score], [Tags], [Title], [ViewCount] 
-                 FROM [dbo].[Posts] 
-                 WHERE [OwnerUserId] IN (SELECT [Id] FROM [Users] WHERE Id < ?)                 
-                    """;
+                 [CreationDate], [FavoriteCount], [LastActivityDate], [LastEditDate], [LastEditorDisplayName], [LastEditorUserId],
+                 [OwnerUserId], [ParentId], [PostTypeId], [Score], [Tags], [Title], [ViewCount]
+                 FROM [dbo].[Posts]
+                 WHERE [OwnerUserId] IN (SELECT [Id] FROM [Users] WHERE Id < ?)
+                 """;
         rowsIn = exec(Post.class, sql, 1000);
 
         // Posts EXISTS
         sql = """
-                SELECT [Id], [AcceptedAnswerId], [AnswerCount], [Body], [ClosedDate], [CommentCount], [CommunityOwnedDate], 
-                 [CreationDate], [FavoriteCount], [LastActivityDate], [LastEditDate], [LastEditorDisplayName], [LastEditorUserId], 
-                 [OwnerUserId], [ParentId], [PostTypeId], [Score], [Tags], [Title], [ViewCount] 
-                 FROM [dbo].[Posts] 
-                 WHERE EXISTS (SELECT [Id] FROM [Users] WHERE Id < ?  AND [Users].[Id] = [Posts].[OwnerUserId])                 
-                    """;
+                SELECT [Id], [AcceptedAnswerId], [AnswerCount], [Body], [ClosedDate], [CommentCount], [CommunityOwnedDate],
+                 [CreationDate], [FavoriteCount], [LastActivityDate], [LastEditDate], [LastEditorDisplayName], [LastEditorUserId],
+                 [OwnerUserId], [ParentId], [PostTypeId], [Score], [Tags], [Title], [ViewCount]
+                 FROM [dbo].[Posts]
+                 WHERE EXISTS (SELECT [Id] FROM [Users] WHERE Id < ?  AND [Users].[Id] = [Posts].[OwnerUserId])
+                 """;
         rowsExists = exec(Post.class, sql, 1000);
         assertEquals(rowsIn, rowsExists);
 
         // Comments IN
         sql = """
-                SELECT [Id], [CreationDate], [PostId], [Score], [Text], [UserId] 
-                 FROM [dbo].[Comments] 
-                 WHERE [PostId] IN (SELECT [Id] FROM [Posts] 
-                    WHERE [OwnerUserId] IN (SELECT [Id] FROM [Users] WHERE Id < ?))                 
+                SELECT [Id], [CreationDate], [PostId], [Score], [Text], [UserId]
+                 FROM [dbo].[Comments]
+                 WHERE [PostId] IN (SELECT [Id] FROM [Posts]
+                    WHERE [OwnerUserId] IN (SELECT [Id] FROM [Users] WHERE Id < ?))
                     """;
         rowsIn = exec(Comment.class, sql, 1000);
 
         // Comments EXISTS
         sql = """
-                SELECT [Id], [CreationDate], [PostId], [Score], [Text], [UserId] 
-                FROM [dbo].[Comments] 
-                 WHERE EXISTS (SELECT [Id] FROM [Posts] 
-                    WHERE EXISTS (SELECT [Id] FROM [Users] WHERE Id < ?  AND [Users].[Id] = [Posts].[OwnerUserId])  AND [Posts].[Id] = [Comments].[PostId])                 
+                SELECT [Id], [CreationDate], [PostId], [Score], [Text], [UserId]
+                FROM [dbo].[Comments]
+                 WHERE EXISTS (SELECT [Id] FROM [Posts]
+                    WHERE EXISTS (SELECT [Id] FROM [Users] WHERE Id < ?  AND [Users].[Id] = [Posts].[OwnerUserId])  AND [Posts].[Id] = [Comments].[PostId])
                     """;
         rowsExists = exec(Comment.class, sql, 1000);
         assertEquals(rowsIn, rowsExists);
 
         // Users IN (from Comments)
         sql = """
-                SELECT [Id], [AboutMe], [Age], [CreationDate], [DisplayName], [DownVotes], [EmailHash], [LastAccessDate], 
-                 [Location], [Reputation], [UpVotes], [Views], [WebsiteUrl], [AccountId] 
-                 FROM [dbo].[Users] 
-                 WHERE [Id] IN (SELECT [UserId] FROM [Comments] 
-                  WHERE [PostId] IN (SELECT [Id] FROM [Posts] 
-                      WHERE [OwnerUserId] IN (SELECT [Id] FROM [Users] WHERE Id < ?)))   
-                  """;
+                SELECT [Id], [AboutMe], [Age], [CreationDate], [DisplayName], [DownVotes], [EmailHash], [LastAccessDate],
+                 [Location], [Reputation], [UpVotes], [Views], [WebsiteUrl], [AccountId]
+                 FROM [dbo].[Users]
+                 WHERE [Id] IN (SELECT [UserId] FROM [Comments]
+                  WHERE [PostId] IN (SELECT [Id] FROM [Posts]
+                      WHERE [OwnerUserId] IN (SELECT [Id] FROM [Users] WHERE Id < ?)))
+                      """;
         rowsIn = exec(User.class, sql, 1000);
 
         // Users EXISTS (from Comments)
         sql = """
-                SELECT [Id], [AboutMe], [Age], [CreationDate], [DisplayName], [DownVotes], [EmailHash], [LastAccessDate], 
-                 [Location], [Reputation], [UpVotes], [Views], [WebsiteUrl], [AccountId] 
-                 FROM [dbo].[Users] 
-                 WHERE EXISTS (SELECT [UserId] FROM [Comments] 
-                    WHERE EXISTS (SELECT [Id] FROM [Posts] 
-                        WHERE EXISTS (SELECT [Id] FROM [Users] WHERE Id < ?  AND [Users].[Id] = [Posts].[OwnerUserId])  AND [Posts].[Id] = [Comments].[PostId])  AND [Comments].[UserId] = [Users].[Id]) 
-                """;
+                SELECT [Id], [AboutMe], [Age], [CreationDate], [DisplayName], [DownVotes], [EmailHash], [LastAccessDate],
+                 [Location], [Reputation], [UpVotes], [Views], [WebsiteUrl], [AccountId]
+                 FROM [dbo].[Users]
+                 WHERE EXISTS (SELECT [UserId] FROM [Comments]
+                    WHERE EXISTS (SELECT [Id] FROM [Posts]
+                        WHERE EXISTS (SELECT [Id] FROM [Users] WHERE Id < ?  AND [Users].[Id] = [Posts].[OwnerUserId])
+                        AND [Posts].[Id] = [Comments].[PostId])  AND [Comments].[UserId] = [Users].[Id])
+                        """;
         rowsExists = exec(User.class, sql, 1000);
         assertEquals(rowsIn, rowsExists);
 
         // Comment for post user IN
         sql = """
-                SELECT [Id], [CreationDate], [PostId], [Score], [Text], [UserId] 
-                 FROM [dbo].[Comments] 
-                 WHERE [PostId] IN (SELECT [Id] FROM [Posts] 
-                    WHERE [OwnerUserId] IN (SELECT [Id] FROM [Users] WHERE Id < ?))  AND [UserId] IN (SELECT [OwnerUserId] FROM [Posts] 
-                        WHERE [OwnerUserId] IN (SELECT [Id] FROM [Users] WHERE Id < ?))        
-                """;
+                SELECT [Id], [CreationDate], [PostId], [Score], [Text], [UserId]
+                 FROM [dbo].[Comments]
+                 WHERE [PostId] IN (SELECT [Id] FROM [Posts]
+                    WHERE [OwnerUserId] IN (SELECT [Id] FROM [Users] WHERE Id < ?))  AND [UserId] IN (SELECT [OwnerUserId] FROM [Posts]
+                        WHERE [OwnerUserId] IN (SELECT [Id] FROM [Users] WHERE Id < ?))
+                        """;
         rowsIn = exec(Comment.class, sql, 1000, 1000);
 
         // comment for post user EXISTS
         sql = """
-                SELECT [Id], [CreationDate], [PostId], [Score], [Text], [UserId] 
-                 FROM [dbo].[Comments] 
-                 WHERE EXISTS (SELECT [Id],[OwnerUserId] FROM [Posts] 
-                    WHERE EXISTS (SELECT [Id] FROM [Users] WHERE Id < ?  AND [Users].[Id] = [Posts].[OwnerUserId])  AND [Posts].[Id] = [Comments].[PostId] AND [Posts].[OwnerUserId] = [Comments].[UserId])                 
+                SELECT [Id], [CreationDate], [PostId], [Score], [Text], [UserId]
+                 FROM [dbo].[Comments]
+                 WHERE EXISTS (SELECT [Id],[OwnerUserId] FROM [Posts]
+                    WHERE EXISTS (SELECT [Id] FROM [Users] WHERE Id < ?  AND [Users].[Id] = [Posts].[OwnerUserId])  AND [Posts].[Id] = [Comments].[PostId] AND [Posts].[OwnerUserId] = [Comments].[UserId])
                 """;
         rowsExists = exec(Comment.class, sql, 1000);
         // these are NOT equal. IN with multiple rows returns EXTRA WE WONT WANT
@@ -237,25 +238,26 @@ public class TestStackOverflow extends TestCase {
 
         //  user comment for post user IN
         sql = """
-                SELECT [Id], [AboutMe], [Age], [CreationDate], [DisplayName], [DownVotes], [EmailHash], [LastAccessDate], 
-                 [Location], [Reputation], [UpVotes], [Views], [WebsiteUrl], [AccountId] 
-                 FROM [dbo].[Users] 
-                 WHERE [Id] IN (SELECT [UserId] FROM [Comments] 
-                    WHERE [PostId] IN (SELECT [Id] FROM [Posts] 
-                        WHERE [OwnerUserId] IN (SELECT [Id] FROM [Users] WHERE Id < ?))  AND [UserId] IN (SELECT [OwnerUserId] FROM [Posts] 
-                            WHERE [OwnerUserId] IN (SELECT [Id] FROM [Users] WHERE Id < ?)))                 
-                    """;
+                SELECT [Id], [AboutMe], [Age], [CreationDate], [DisplayName], [DownVotes], [EmailHash], [LastAccessDate],
+                 [Location], [Reputation], [UpVotes], [Views], [WebsiteUrl], [AccountId]
+                 FROM [dbo].[Users]
+                 WHERE [Id] IN (SELECT [UserId] FROM [Comments]
+                    WHERE [PostId] IN (SELECT [Id] FROM [Posts]
+                        WHERE [OwnerUserId] IN (SELECT [Id] FROM [Users] WHERE Id < ?))  AND [UserId] IN (SELECT [OwnerUserId] FROM [Posts]
+                            WHERE [OwnerUserId] IN (SELECT [Id] FROM [Users] WHERE Id < ?)))
+                            """;
         rowsIn = exec(User.class, sql, 1000, 1000);
 
         //  user comment for post user EXISTS
         sql = """
-                SELECT [Id], [AboutMe], [Age], [CreationDate], [DisplayName], [DownVotes], [EmailHash], [LastAccessDate], 
-                 [Location], [Reputation], [UpVotes], [Views], [WebsiteUrl], [AccountId] 
-                 FROM [dbo].[Users] 
-                 WHERE EXISTS (SELECT [UserId] FROM [Comments] 
-                    WHERE EXISTS (SELECT [Id],[OwnerUserId] FROM [Posts] 
-                        WHERE EXISTS (SELECT [Id] FROM [Users] WHERE Id < ?  AND [Users].[Id] = [Posts].[OwnerUserId])  AND [Posts].[Id] = [Comments].[PostId] AND [Posts].[OwnerUserId] = [Comments].[UserId])  AND [Comments].[UserId] = [Users].[Id])                 
-                    """;
+                SELECT [Id], [AboutMe], [Age], [CreationDate], [DisplayName], [DownVotes], [EmailHash], [LastAccessDate],
+                 [Location], [Reputation], [UpVotes], [Views], [WebsiteUrl], [AccountId]
+                 FROM [dbo].[Users]
+                 WHERE EXISTS (SELECT [UserId] FROM [Comments]
+                    WHERE EXISTS (SELECT [Id],[OwnerUserId] FROM [Posts]
+                        WHERE EXISTS (SELECT [Id] FROM [Users] WHERE Id < ?  AND [Users].[Id] = [Posts].[OwnerUserId])
+                        AND [Posts].[Id] = [Comments].[PostId] AND [Posts].[OwnerUserId] = [Comments].[UserId])  AND [Comments].[UserId] = [Users].[Id])
+                        """;
         rowsExists = exec(User.class, sql, 1000);
         // Same with comment query
         assertTrue(rowsIn > rowsExists);