+++ test/net/sf/persism/dao/ContactFail.java
@@ -59,9 +59,9 @@ public final class ContactFail {
         this.firstname = firstname;
     }
 
-    public String getLastname() {
-        return lastname;
-    }
+//    public String getLastname() {
+//        return lastname;
+//    }
 
     public void setLastname(String lastname) {
         this.lastname = lastname;